add ReadObjectProp method
Intended to be used for header reading without
the need to read the entire file.

It does not work like that on JS as
localStorage has no partial read API,
but it will still copy the data to the
provided preallocated buffer.

The alternative would be to return a io.ReadCloser
and give an Open-like method, but that
would expose more operations than necessary.
To keep this library simple and portable,
I decided to go with a more conservative API
that is more limiting.